requirements.txt
docker==2.0.2
python-etcd==0.4.4
